libft-typing fixed

+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: mbenomar <mbenomar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 19:04:24 by mbenomar          #+#    #+#             */
-/*   Updated: 2024/10/24 08:50:45 by mbenomar         ###   ########.fr       */
+/*   Updated: 2024/10/24 20:25:32 by mbenomar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && s1[i] == s2[i])
